(home-di) add BaseViewController, HomeDIContainer
