+++ soma-weather-viewer-ios/soma-weather-viewer-ios/app/AppFlowCoordinator.swift
@@ -22,7 +22,10 @@ final class AppFlowCoordinator {
     }
     
     func start() {
-        let homeViewController = appDIContainer.makeHomeViewController()
+        // In App Flow we can check if user needs to login, if yes we would run login flow
+        let homeDIContainer = appDIContainer.makeHomeDIContainer()
+        let homeViewController = homeDIContainer.makeHomeViewController()
+        
         navigationController.pushViewController(homeViewController, animated: false)
     }
